Updating component import to restructure payload metrics display
