+++ pages/_app.js
@@ -1,10 +1,11 @@
 import App from 'next/app';
 import Head from 'next/head';
 import GlobalStyle from '@catho/quantum/GlobalStyle';
-import { Container } from '@catho/quantum';
+import Container from '@catho/quantum/Grid/sub-components/Container';
 import ScrollUpButton from '../components/ScrollUpButton';
 import '../static/style.css';
-import { Footer, Header } from '../components';
+import Footer from '../components/Footer';
+import Header from '../components/Header';
 
 export default class HomePage extends App {
   render() {
