+++ pages/render.js
@@ -2,7 +2,7 @@ import Container from '@catho/quantum/Grid/sub-components/Container';
 import Col from '@catho/quantum/Grid/sub-components/Col';
 import Row from '@catho/quantum/Grid/sub-components/Row';
 import HeaderResponsive from '@catho-private/catho-components/HeaderResponsive';
-import { FooterTagsList } from '../components';
+import FooterTagsList from '../components/FooterTagsList';
 import Header from './shared/Header';
 
 export default function RenderPage() {
